+++ product-service/src/test/java/productservice/entities/PublisherJpaTest.java
@@ -4,7 +4,7 @@ import jakarta.persistence.EntityManager;
 import jakarta.validation.ConstraintViolationException;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
+import org.springframework.boot.data.jpa.test.autoconfigure.DataJpaTest;
 import productservice.repositories.PublisherRepository;
 
 import java.util.Optional;
@@ -133,7 +133,7 @@ class PublisherJpaTest {
         entityManager.flush();
 
         assertNotNull(saved.getVersion());
-        Long initialVersion = saved.getVersion();
+        Integer initialVersion = saved.getVersion();
 
         saved.setName("Updated Lock Test Publisher");
         Publisher updated = publisherRepository.save(saved);
